my_mastermind.c
my_mastermind will be an implementation of the famous game. Mastermind is a game composed of 8 pieces of different colors. A secret code is then composed of 4 distinct pieces. The player has 10 attempts to find the secret code. Pieces will be '0' '1' '2' '3' '4' '5' '6' '7'. If the player finds the code, he wins, and the game stops. You must read the player's input from the standard input. Your program will also receive the following parameters: -c [CODE]: specifies the secret code. If no code is specified, a random code will be generated.